Add file name stuff
`sudo apt install net-tools` works for ubuntu, so it's not "only" archlinux/manjaro. Kinda confused me-

In fact, this whole installation thing confused a few of my friends who took a few tries to install stuff. So I've added the names of the packages so even the most basic morons can install it without somebody else going "No no, it's firejail - as in, without a capital letter.

Need to add the other packages for the other required stuff-